simdhash-related cleanups
