+++ src/coreclr/interpreter/interpreter.h
@@ -25,8 +25,9 @@
 
 #define ALIGN_UP_TO(val,align) ((((size_t)val) + (size_t)((align) - 1)) & (~((size_t)(align - 1))))
 
-#ifdef DEBUG
 extern "C" void assertAbort(const char* why, const char* file, unsigned line);
+
+#ifdef DEBUG
 #undef assert
 #define assert(p) (void)((p) || (assertAbort(#p, __FILE__, __LINE__), 0))
 #endif // DEBUG
